requirements.txt
torch>=2.0.0
torchaudio
librosa
numpy
pandas
pyyaml
scipy
scikit-learn
matplotlib
tqdm
pysoundfile
audiomentations
wandb
tensorboard
shap
lime
scikit-fuzzy
ffmpeg-python
